+++ lib/hexapdf/type/annotations/polygon.rb
@@ -34,34 +34,31 @@
 # commercial licenses are available at <https://gettalong.at/hexapdf/>.
 #++
 
+require 'hexapdf/type/annotations'
+
 module HexaPDF
   module Type
-
-    autoload(:Annotation, 'hexapdf/type/annotation')
-
-    # Namespace module for all PDF annotation dictionary types.
-    #
-    # See: PDF2.0 s12.5.6, Annotation
     module Annotations
 
-      autoload(:MarkupAnnotation, 'hexapdf/type/annotations/markup_annotation')
-      autoload(:Text, 'hexapdf/type/annotations/text')
-      autoload(:Link, 'hexapdf/type/annotations/link')
-      autoload(:Widget, 'hexapdf/type/annotations/widget')
-      autoload(:BorderStyling, 'hexapdf/type/annotations/border_styling')
-      autoload(:Line, 'hexapdf/type/annotations/line')
-      autoload(:AppearanceGenerator, 'hexapdf/type/annotations/appearance_generator')
-      autoload(:BorderEffect, 'hexapdf/type/annotations/border_effect')
-      autoload(:InteriorColor, 'hexapdf/type/annotations/interior_color')
-      autoload(:SquareCircle, 'hexapdf/type/annotations/square_circle')
-      autoload(:Square, 'hexapdf/type/annotations/square')
-      autoload(:Circle, 'hexapdf/type/annotations/circle')
-      autoload(:LineEndingStyling, 'hexapdf/type/annotations/line_ending_styling')
-      autoload(:PolygonPolyline, 'hexapdf/type/annotations/polygon_polyline')
-      autoload(:Polygon, 'hexapdf/type/annotations/polygon')
-      autoload(:Polyline, 'hexapdf/type/annotations/polyline')
+      # A polygon annotation displays a closed polygon inside the annotation rectangle.
+      #
+      # Also see PolygonPolyline for more information.
+      #
+      # Example:
+      #
+      #   #>pdf-small
+      #   doc.annotations.create_polygon(doc.pages[0], 20, 20, 30, 70, 80, 60, 40, 30).
+      #     border_style(color: "hp-blue", width: 2, style: [3, 1]).
+      #     interior_color("hp-orange").
+      #     regenerate_appearance
+      #
+      # See: PDF2.0 s12.5.6.9, HexaPDF::Type::Annotations::PolygonPolyline
+      class Polygon < PolygonPolyline
+
+        define_field :Subtype, type: Symbol, required: true, default: :Polygon
+
+      end
 
     end
-
   end
 end